Made betty compliant and compile

+++ 0x1E-search_algorithms/0-linear.c
@@ -8,7 +8,7 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t indx;
+	unsigned int indx;
 
 	if (array == NULL || size == 0)
 		return (-1);
